+++ R/check_tbl_spl_compound_tid.R
@@ -1,8 +1,12 @@
 #' Check model output data tbl samples contain single unique values for each
 #' compound task ID within individual samples
+#'
 #' @param tbl a tibble/data.frame of the contents of the file being validated. Column types must **all be character**.
 #' @inherit check_tbl_colnames params
 #' @inherit check_tbl_colnames return
+#' @param compound_taskid_set a list of `compound_taskid_set`s (characters vector of compound task IDs),
+#' one for each modeling task. Used to override the compound task ID set in the config file,
+#' for example, when validating coarser samples.
 #' @details Output of the check includes an `errors` element, a list of items,
 #' one for each sample failing validation, with the following structure:
 #' - `mt_id`: Index identifying the config modeling task the sample is associated with.
@@ -12,16 +16,24 @@
 #' See [hubverse documentation on samples](https://hubverse.io/en/latest/user-guide/sample-output-type.html)
 #' for more details.
 #' @export
-check_tbl_spl_compound_tid <- function(tbl, round_id, file_path, hub_path) {
+check_tbl_spl_compound_tid <- function(tbl, round_id, file_path, hub_path,
+                                       compound_taskid_set = NULL) {
+  if (!is.null(compound_taskid_set) && isTRUE(is.na(compound_taskid_set))) {
+    cli::cli_abort("Valid {.var compound_taskid_set} must be provided.")
+  }
   config_tasks <- hubUtils::read_config(hub_path, "tasks")
+  if (is.null(compound_taskid_set)) {
+    compound_taskid_set <- get_round_compound_task_ids(
+      config_tasks,
+      round_id
+    )
+  }
 
   if (isFALSE(has_spls_tbl(tbl)) || isFALSE(hubUtils::is_v3_config(config_tasks))) {
     return(skip_v3_spl_check(file_path))
   }
 
-  hash_tbl <- spl_hash_tbl(tbl, round_id, config_tasks)
-  # TODO: Currently, samples must strictly match the compound task ID set expectations
-  # and cannot handle coarser-grained compound task ID sets.
+  hash_tbl <- spl_hash_tbl(tbl, round_id, config_tasks, compound_taskid_set)
   n_tbl <- hash_tbl[hash_tbl$n_compound_idx > 1L, ]
 
   check <- nrow(n_tbl) == 0L
@@ -31,7 +43,7 @@ check_tbl_spl_compound_tid <- function(tbl, round_id, file_path, hub_path) {
     errors <- NULL
   } else {
     errors <- comptid_mismatch(
-      n_tbl, tbl, config_tasks, round_id
+      n_tbl, tbl, config_tasks, round_id, compound_taskid_set
     )
     output_type_ids <- purrr::map(errors, ~ .x$output_type_id) %>% # nolint: object_usage_linter
       purrr::flatten_chr() %>%
@@ -57,13 +69,13 @@ check_tbl_spl_compound_tid <- function(tbl, round_id, file_path, hub_path) {
   )
 }
 
-comptid_mismatch <- function(n_tbl, tbl, config_tasks, round_id) {
+comptid_mismatch <- function(n_tbl, tbl, config_tasks, round_id, compound_taskid_set) {
   tbl <- tbl[tbl$output_type == "sample", ]
   purrr::map(
     seq_along(n_tbl$output_type_id),
     ~ {
       x <- n_tbl[.x, ]
-      compound_taskids <- get_round_compound_task_ids(config_tasks, round_id)[[x$mt_id]]
+      compound_taskids <- compound_taskid_set[[x$mt_id]]
       spl <- tbl[tbl$output_type_id == x$output_type_id, compound_taskids] %>%
         unique()