+++ R/validate_model_data.R
@@ -186,12 +186,26 @@ validate_model_data <- function(hub_path, file_path, round_id_col = NULL,
 
   # -- v3 sample checks ----
   if (hubUtils::is_v3_hub(hub_path)) {
+    checks$spl_compound_taskid_set <- try_check(
+      check_tbl_spl_compound_taskid_set(
+        tbl_chr,
+        round_id = round_id,
+        file_path = file_path,
+        hub_path = hub_path
+      ), file_path
+    )
+    if (is_any_error(checks$spl_compound_taskid_set)) {
+      return(checks)
+    } else {
+      compound_taskid_set <- checks$spl_compound_taskid_set$compound_taskid_set
+    }
     checks$spl_compound_tid <- try_check(
       check_tbl_spl_compound_tid(
         tbl_chr,
         round_id = round_id,
         file_path = file_path,
-        hub_path = hub_path
+        hub_path = hub_path,
+        compound_taskid_set = compound_taskid_set
       ), file_path
     )
     if (is_any_error(checks$spl_compound_tid)) {
@@ -202,7 +216,8 @@ validate_model_data <- function(hub_path, file_path, round_id_col = NULL,
         tbl_chr,
         round_id = round_id,
         file_path = file_path,
-        hub_path = hub_path
+        hub_path = hub_path,
+        compound_taskid_set = compound_taskid_set
       ), file_path
     )
     if (is_any_error(checks$spl_non_compound_tid)) {
@@ -213,7 +228,8 @@ validate_model_data <- function(hub_path, file_path, round_id_col = NULL,
         tbl_chr,
         round_id = round_id,
         file_path = file_path,
-        hub_path = hub_path
+        hub_path = hub_path,
+        compound_taskid_set = compound_taskid_set
       ), file_path
     )
   }